Add functionality of SQS Queue Lambda to Fetch and Queue Document IDs Lambda
The SQS Queue Lambda was initially separate from the Fetch Document IDs Lambda.
This commit combines the two lambda into two as the results from the Fetch Documents Lambda is the input of the PushIdsToSqs Usecase.

This commit adds a service configuration for initialising the SQS client passed into the SQS Gateway.