Add S3 endpoint to view the document
This adds an endpoint that redirects to the document in S3.
The S3 URL is generated when a GET request is made to `/api/documents/{id}/view` where id is the id of the document selected for viewing.

The generated URL is a pre-signed URL that has an expiration time of 5 minutes, within which the user can view the document. This time frame was chosen based on examples from the AWS documentation and can be changed if there is a better time frame for URL expiration.