Part of the CTax Enquiry Form template parser.
Added some changes to the Letter Template object to accomodate the possibility of the 'actual' letter head (as opposed from the the addresses one from before).

Reason for that was because the address ones would always be on the top of whatever other content, so the true head needed separating to be on top. This is reflected in helper function changes of parser unrelated files.