Changed the method of parsing table cell layout.
For some unclear reason, due to the Environment differences, the code on the cloud would flip the positions of the second to last two table no.3 cells.
(when compared to the proper order given by the code running locally).

To fix this issue, the C# way of getting rid of rowspans, and reordering the table had to go, and CSS had to be used instead.

In addition to that this commit also fixes the cell border allignment across all tables of the form.